requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
pandas==2.1.3
json5==0.9.14
python-dotenv==1.0.0
urllib3==2.1.0
fake-useragent==1.4.0
webdriver-manager==4.0.2

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/LLM
google-generativeai==0.8.3

# RAG and Vector Database
chromadb==1.3.4
numpy==1.24.3

